+++ include/GafferSceneUI/SelectionTool.h
@@ -39,6 +39,8 @@
 #include "GafferSceneUI/Export.h"
 #include "GafferSceneUI/TypeIds.h"
 
+#include "GafferScene/ScenePlug.h"
+
 #include "GafferUI/DragDropEvent.h"
 #include "GafferUI/Tool.h"
 
@@ -63,6 +65,19 @@ class GAFFERSCENEUI_API SelectionTool : public GafferUI::Tool
 
 		Gaffer::StringPlug *selectModePlug();
 		const Gaffer::StringPlug *selectModePlug() const;
+
+		using SelectFunction = std::function<GafferScene::ScenePlug::ScenePath(
+			const GafferScene::ScenePlug *,
+			const GafferScene::ScenePlug::ScenePath &
+		)>;
+		// Registers a select mode identified by `name`. `function` must accept
+		// the scene from which a selection will be made and the `ScenePath` the user
+		// initially selected. It returns the `ScenePath` to use as the actual selection.
+		static void registerSelectMode( const std::string &name, SelectFunction function );
+		// Returns the names of registered modes, in the order they were registered.
+		// The "/Standard" mode will always be first.
+		static std::vector<std::string> registeredSelectModes();
+		static void deregisterSelectMode( const std::string &mode );
 
 	private :