+++ python/GafferSceneUITest/SelectionToolTest.py
@@ -1,6 +1,6 @@
 ##########################################################################
 #
-#  Copyright (c) 2013, Image Engine Design Inc. All rights reserved.
+#  Copyright (c) 2024, Cinesite VFX Ltd. All rights reserved.
 #
 #  Redistribution and use in source and binary forms, with or without
 #  modification, are permitted provided that the following conditions are
@@ -34,35 +34,34 @@
 #
 ##########################################################################
 
-from .SceneViewTest import SceneViewTest
-from .ShaderAssignmentUITest import ShaderAssignmentUITest
-from .StandardGraphLayoutTest import StandardGraphLayoutTest
-from .SceneGadgetTest import SceneGadgetTest
-from .SceneInspectorTest import SceneInspectorTest
-from .HierarchyViewTest import HierarchyViewTest
-from .DocumentationTest import DocumentationTest
-from .ShaderViewTest import ShaderViewTest
-from .ShaderUITest import ShaderUITest
-from .TranslateToolTest import TranslateToolTest
-from .ScaleToolTest import ScaleToolTest
-from .RotateToolTest import RotateToolTest
-from .ContextAlgoTest import ContextAlgoTest
-from .CameraToolTest import CameraToolTest
-from .VisualiserTest import VisualiserTest
-from .TransformToolTest import TransformToolTest
-from .CropWindowToolTest import CropWindowToolTest
-from .NodeUITest import NodeUITest
-from .ParameterInspectorTest import ParameterInspectorTest
-from .AttributeInspectorTest import AttributeInspectorTest
-from .HistoryPathTest import HistoryPathTest
-from .LightEditorTest import LightEditorTest
-from .SetMembershipInspectorTest import SetMembershipInspectorTest
-from .SetEditorTest import SetEditorTest
-from .LightToolTest import LightToolTest
-from .OptionInspectorTest import OptionInspectorTest
-from .LightPositionToolTest import LightPositionToolTest
-from .RenderPassEditorTest import RenderPassEditorTest
-from .SelectionToolTest import SelectionToolTest
+import unittest
 
-if __name__ == "__main__":
+import GafferUITest
+import GafferSceneUI
+
+class SelectionToolTest( GafferUITest.TestCase ) :
+
+	def modifierFunction( scene, path ) :
+
+		return path
+
+	def testRegisterSelectMode( self ) :
+
+		GafferSceneUI.SelectionTool.registerSelectMode( "testModifier", self.modifierFunction )
+		GafferSceneUI.SelectionTool.registerSelectMode( "testModifier2", self.modifierFunction )
+
+		modifiers = GafferSceneUI.SelectionTool.registeredSelectModes()
+		self.assertEqual( len( modifiers ), 3 )
+
+		self.assertEqual( modifiers, [ "/Standard", "testModifier", "testModifier2" ] )
+
+	def tearDown( self ) :
+
+		GafferUITest.TestCase.tearDown( self )
+
+		GafferSceneUI.SelectionTool.deregisterSelectMode( "testModifier" )
+		GafferSceneUI.SelectionTool.deregisterSelectMode( "testModifier2" )
+
+
+if __name__ == "__main__" :
 	unittest.main()